+++ fabric/CHANGELOG.md
@@ -1,6 +1,10 @@
 # Nexus API Changelog (Fabric)
 
-# v1.0.6 (1.20.1)
+# v1.0.7 (1.20.1)
 
 ## Bug Fixes
 
+
+## New Features
+
+- Added new BlockPropertyWrapper template for leaves in BlockPropertyWrapperTemplates.